added docker-compose config validation
